Disable the github actions CI
They're doing SOMETHING with UDP packets, and I can't get to the bottom of it,
and it causes all UDP tests to time out. My guess is that they limit us to a
single UDP listener, and since you need to create a listener to send a UDP
packet, that's a non-starter for us.

I'll swap to another CI in the future.