Stop using std logger, and instead depend on an injected one.
Directly accessing the std logger, but expecting the user to configure it is really jank,
since it depends on std versions lining up exactly, which won't always happen. So instead,
we define a reasonable subset of the logger api, and just have the implementation injected.

This lets us have our fancy logger, but gives the user full control over it. Maybe deno
will get built-in dependency injection at some point, or a better pattern will emerge. But
this is good enough for now.